onecuponetree/requirements_no_gdal.txt
Django==4.2.7
django-environ==0.11.2
django-leaflet==0.28.3
django-modeltranslation==0.19.2
Pillow  # Updated version that supports Python 3.11
psycopg2-binary
whitenoise==6.5.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-allauth==0.58.2
django-countries==7.5.1
django-filter==23.3
django-tables2==2.6.0
django-widget-tweaks==1.5.0
